Set CTest's working directory to `tests/`
The `gtest_discover_tests()` command in CMake defaults to using
relative paths based on the build directory when running the tests. This
can cause issues if developers build multiple versions of the project,
for example with different configuration flags, as the tests will fail
to locate the test files correctly.

To decouple the build process from the test execution, this commit sets
the working directory for CTest to the tests/ subdirectory. By doing so,
the relative paths used by the discovered tests will correctly resolve
to the test files, regardless of the build directory structure or
configuration used.