Add a class for generating type pairs for testing
Sometimes we want to create a test case testing with all combinations of
the supported data types. There is a test checking if the value of
each data type can be filled into the `Storage` instance storing the
values in every supported data type. One way to make this kind of test
case but not to duplicate the test cases for each combination of types
is to use gtest's typed-test feature which needs us to list all
combinations of types.

```cpp
using MyTypes = ::testing::Types<std::pair<int, double>,
                                 std::pair<int, bool>,
                                 std::pair<double, double>,
                                 std::pair<double, bool>>;
TYPED_TEST_SUITE(MyFixture, MyTypes);
TYPED_TEST(TestSuiteName, TestName) {
  using FirstType = TypeParam::first_type;
  using SecondType = TypeParam::second_type;
  //...
}
```

In the example `MyTypes` contains the combination of two groups of
types. The first group is `int` and `double`, and the second group is
`double` and `bool`. If there are ten supported types, there will be one
hundred pairs in `MyTypes`. Writing that kind of code is exhausting,
and the code will be hard to maintain.

This change provides a tool for generating all combinations of the given
testing types and makes writing tests more easily.