Reformatted with git-clang-format
Also fixed a bogus change to src/UniTensor.cpp.  I didn't intend to include src/UniTensor.cpp in
the previous commit, it was an accident and the change was some spurious debugging stuff. That
is now reverted, and took the opportunity for some simple cleanups of the error messages, to
make them more uniform with Tensor.cpp