Remove `-fopenmp` flag when only enabling MKL
We don't guard `#pragam omp...` statements now, so they take effects
when `-fopenmp` compiler flag is set. `set()` function only affect the
targets defined in the same scope (the same directory) and the
subdirectories after the line of `set()`. Setting `-fopenmp` flag when
the user only enable `USE_MKL` but not enable `USE_OMP` makes the user
suprised.

By default, MKL links to their own openmp (not gomp) for threading. The
build process has no problem with no setting `-fopenmp` compiler flag
when using MKL as the provider of BLAS and LAPACK. The above claim is
true for the MKL version later then 2023.2.0. I am not sure if it is
also true for the older version of MKL.