Implement Storage::release()
`Storage::release()` serves for the function converting Storage and
Tensor to the numpy array. The reason for implementing this method is
that we are encapsulating Storage and the relative classes, which mean
we don't want other classes or functions to directly access the member
variables.

This is just a temporary solution and will be removed when we use
std::vector as the underlying storage because it is impossible to make
std::vector drops its ownership of the underlying storage.