Encapsulate Storage_base class
With the current class definition, developers are able to change the
variable storing the size of the storage without allocating the
underlying storage, which potentially causes memory leaks.

To ensure there will be no surprising behavior caused by directly
accessing the variables of Storage_base, all variables of Stroage_base
are moved to its subclass, StorageImplementation<DType>, and are set as
private. All functions outside of StorageImplementation<DType> should
access the value of the variable via the getters.

The next step in refactoring the components related to Storage is to
remove the Storage_base class.