Make tests for Det() buildable
The test cases are implemented with some problems, so they are expected
to fail. The problems below need to be fixed in the future:

1. Both CPU and GPU algorithms are frustrating in handling the boolean
type.
2. The tolerance value for verifying the result is wrong. `abs(result_value) / 1e6`
may be a choice.
3. The seed for random generator is fixed, so the randomized testing
tensors are identical across sessions. If this is intended design, we
should better to explicitly code the values in the tensors out.
4. The tests for GPU don't run on GPU.
5. The test names of GPU test and CPU test collide with each other.