Define all options and variables in CMakeLists.txt
Add the options and variables defined in Install.sh but not in
CMakeLists.txt to CMakeLists.txt. The dependencies between the options
and variables are maintained with the CMake functions.

Install.sh will be removed after applying CMakePresets.json to define
the default compile options and variables for each preset.