Replace Install.sh with CMakePresets.json
CMakePresets.json is a recommend method to list all settings for CMake.

`CUTENSOR_ROOT` and `CUQUANTUM_ROOT` are not explicitly added into
CMakePresets.json because vscode-cmake-tools, an vscode extension for
CMake, doesn't pass environment variables to presets. The users of
the extension will have no intuitive way to pass the value of
`CUTENSOR_ROOT` and `CUQUANTUM_ROOT` for configuration. As an
alternative we show an error message to the user when they doesn't set
`CUTENSOR_ROOT` or `CUQUANTUM_ROOT` in the environment variables.