Propagate `-fopenmp` linking flag when using HPTT
When enabling HPTT and using OpenBLAS as the BLAS vendor, we didn't
propagate `-fopenmp` flag. As a result, the executables linking with
Cytnx static library will fail to find the definition of OpenMP, which
is used by HPTT.

The GitHub flow only building Cytnx with enabling HPTT and using MKL, so
this issue wasn't caught by the CI before.