Fix Mamba error when running GitHub actions
Drop boa and mamba because they almost don't speed the build process
up comparing with using the built-in conda-build now. Conda uses
libmamba as the default solver since 23.10, which made conda comparable
to mamba. Additaionally, boa was archived in 2024 Dec. The replacement
of boa is rattler-build. rattler-build may reduce some overheads of
conda-build. It's worth to try in the future.

Downgrade the Python version to 3.10 to support older Python version
in the future. Alought the oldest Python version under security
maintaince is 3.9 now, the support will be terminated in 2025 Oct. To
avoid changing the Python version too freqently, downgrade the version
to 3.10 instead of 3.9.