Improves Conda build workflow with caching
Adds caching for the build results with Ccache to improve speed of the
workflow for PRs. This reduces the time spent on building the project
for each version of Python and for each run. With caching with Ccache,
the build time reduces about 85%, from 600 senconds to 90 seconds.

For each new pull request, the cache is loaded from the main branch. For
each new commit in the pull request, the cache is loaded from the
previous commit in the pull request.