Refactors Conda build workflows for cross-platform
Consolidates Conda build workflows for Linux and macOS into single
workflows using a matrix strategy.

This change simplifies the CI configuration and enables testing on both
platforms.

The separate macOS workflows are removed, and a new environment variable
is introduced to specify the correct conda build recipe based on the
operating system.