Update CMake configuration to conditionally set BLA_VENDOR based on MKL interface
This commit modifies the CytnxBKNDCMakeLists.cmake file to conditionally set the BLA_VENDOR variable based on the MKL_INTERFACE value. If the MKL_INTERFACE is set to "ilp64", BLA_VENDOR is set to Intel10_64ilp; otherwise, it defaults to Intel10_64lp. This change enhances the flexibility of the CMake configuration for different MKL interface requirements.